backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
asyncpg
